apps/audio-engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vozazi-audio-engine"
version = "1.0.0"
description = "VOZAZI Audio Processing Engine"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "torchaudio>=2.2.0",
    "torch>=2.2.0",
    "torchcrepe>=0.0.20",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "scipy>=1.12.0",
    "numpy>=1.26.0",
    "essentia-tensorflow>=2.1b6",
    "websockets>=12.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "pgvector>=0.2.0",
    "sentence-transformers>=2.4.0",
    "openai>=1.14.0",
    "anthropic>=0.18.0",
    "aiohttp>=3.9.0",
    "redis>=5.0.0",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.0",
    "prometheus-client>=0.20.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-env>=1.1.0",
    "pytest-timeout>=2.2.0",
    "asgi-lifespan>=2.1.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    # Linting & Formatting
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
    # Security
    "bandit>=1.7.0",
    "safety>=3.0.0",
    "pip-audit>=2.7.0",
    # Type stubs
    "types-redis>=4.6.0",
    "types-aiofiles>=23.2.0",
    "types-python-dateutil>=2.8.0",
    "types-requests>=2.31.0",
    # Development
    "ipython>=8.22.0",
    "jupyter>=1.0.0",
]

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
    "PERF",   # perflint
    "PYI",    # flake8-pyi
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLOT",   # flake8-slots
    "PGH",    # pygrep-hooks
    "FLY",    # flynt
    "C90",    # mccabe complexity
    "PL",     # pylint
    "TRY",    # tryceratops
    "FAST",   # fastapi
    "AIR",    # airflow
    "DOC",    # pydocstyle
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "TRY003", # avoid specifying long messages outside exception class
    "DOC",    # ignore docstring requirements for now
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # unused imports
"tests/**/*.py" = ["S101", "ARG", "PLR2004"]  # assert, unused args, magic numbers
"app/api/**/*.py" = ["B008"]  # function calls in defaults for FastAPI
"conftest.py" = ["F401", "F403"]  # fixtures

[tool.ruff.mccabe]
max-complexity = 12

[tool.ruff.pylint]
allow-magic-value-types = ["int", "str", "float", "bytes"]
max-args = 8
max-branches = 15
max-returns = 8
max-statements = 60

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = """
    -v 
    --cov=app 
    --cov-report=term-missing 
    --cov-report=html:coverage/html 
    --cov-report=xml:coverage/coverage.xml
    --cov-fail-under=80
    --strict-markers
    --strict-config
    --tb=short
    --maxfail=5
"""
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "unit: marks tests as unit tests",
    "requires_db: marks tests as requiring database",
    "requires_redis: marks tests as requiring redis",
    "requires_openai: marks tests as requiring OpenAI API",
    "requires_anthropic: marks tests as requiring Anthropic API",
    "requires_r2: marks tests as requiring R2 storage",
]
env = [
    "DATABASE_URL=postgresql://postgres:postgres@localhost:5432/vozazi_test",
    "REDIS_URL=redis://localhost:6379",
    "LOG_LEVEL=DEBUG",
    "TESTING=true",
]
timeout = 300
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/main.py",
    "*/config.py",
    "*/conftest.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "class .*Protocol.*",
]
fail_under = 80
show_missing = true
skip_covered = true

[tool.bandit]
targets = ["app"]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]  # assert_used, paramiko_calls

[tool.commitlint]
extends = ["@commitlint/config-conventional"]
